+++ packages/extension-bold/CHANGELOG.md
@@ -1,5 +1,17 @@
 # Change Log
 
+## 3.26.1
+
+### Patch Changes
+
+- @tiptap/core@3.26.1
+
+## 3.26.0
+
+### Patch Changes
+
+- @tiptap/core@3.26.0
+
 ## 3.25.0
 
 ### Patch Changes