+++ packages/extension-collaboration/CHANGELOG.md
@@ -1,5 +1,20 @@
 # Change Log
 
+## 3.26.1
+
+### Patch Changes
+
+- a38c9c0: Bump `@tiptap/y-tiptap` to version ^3.0.5
+  - @tiptap/core@3.26.1
+  - @tiptap/pm@3.26.1
+
+## 3.26.0
+
+### Patch Changes
+
+- @tiptap/core@3.26.0
+- @tiptap/pm@3.26.0
+
 ## 3.25.0
 
 ### Patch Changes