+++ packages/markdown/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @tiptap/markdown
 
+## 3.26.1
+
+### Patch Changes
+
+- @tiptap/core@3.26.1
+- @tiptap/pm@3.26.1
+
+## 3.26.0
+
+### Patch Changes
+
+- @tiptap/core@3.26.0
+- @tiptap/pm@3.26.0
+
 ## 3.25.0
 
 ### Patch Changes