+++ packages/pm/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Change Log
 
+## 3.26.1
+
+## 3.26.0
+
 ## 3.25.0
 
 ### Patch Changes